requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
table2ascii==0.3.0
youtube_dl==2021.12.17
beautifulsoup4==4.11.1
discord==2.0.0
pafy==0.5.5
Pillow==9.2.0